Now when you load a file with error, you don't show messageBox
svn path=/trunk/koffice/; revision=74907

+++ kspread/kspread_undo.h
@@ -37,12 +37,12 @@ class RowLayout;
 
 struct rowSize {
 int rowNumber;
-int rowHeight;
+double rowHeight;
 };
 
 struct columnSize {
 int columnNumber;
-int columnWidth;
+double columnWidth;
 };
 
 struct textOfCell {
